Move classes to their own class files
