+++ src/games/strategy/triplea/delegate/AbstractMoveDelegate.java
@@ -235,11 +235,3 @@ public abstract class AbstractMoveDelegate extends BaseTripleADelegate implement
 }
 
 
-class AbstractMoveExtendedDelegateState implements Serializable {
-  private static final long serialVersionUID = -4072966724295569322L;
-  Serializable superState;
-  // add other variables here:
-  public boolean m_nonCombat;
-  public List<UndoableMove> m_movesToUndo;
-  public MovePerformer m_tempMovePerformer;
-}
