+++ src/games/strategy/triplea/formatter/MyFormatter.java
@@ -318,26 +318,3 @@ public class MyFormatter {
 }
 
 
-class UnitOwner {
-  public UnitType type;
-  public PlayerID owner;
-
-  UnitOwner(final UnitType type, final PlayerID id) {
-    this.type = type;
-    this.owner = id;
-  }
-
-  @Override
-  public boolean equals(final Object o) {
-    if (o == null) {
-      return false;
-    }
-    final UnitOwner other = (UnitOwner) o;
-    return other.type.equals(this.type) && other.owner.equals(this.owner);
-  }
-
-  @Override
-  public int hashCode() {
-    return type.hashCode() ^ owner.hashCode();
-  }
-}
